# ui/packages/tidb-dashboard-client/swagger/spec.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dashboard API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/dashboard/api",
    "paths": {
...
        "/topsql/config": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get Top SQL config",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/topsql.EditableConfig"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Update Top SQL config",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topsql.EditableConfig"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topsql/instances": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get availiable instances",
                "parameters": [
                    {
                        "type": "string",
                        "name": "data_source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "start",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/topsql.InstanceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topsql/summary": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get summaries",
                "parameters": [
                    {
                        "type": "string",
                        "name": "data_source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "group_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "instance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "instance_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "top",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "window",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/topsql.SummaryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/topsql/tikv_network_io_collection": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Get TiKV network IO collection config",
                "operationId": "topsqlGetTiKVNetworkIOCollection",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/topsql.TikvNetworkIoCollectionConfig"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "summary": "Update TiKV network IO collection config",
                "operationId": "topsqlUpdateTiKVNetworkIOCollection",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topsql.UpdateTikvNetworkIoCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/topsql.UpdateTikvNetworkIoCollectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "boolean"
                },
                "full_text": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "topsql.EditableConfig": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                }
            }
        },
        "topsql.InstanceItem": {
            "type": "object",
            "properties": {
                "instance": {
                    "type": "string"
                },
                "instance_type": {
                    "type": "string"
                }
            }
        },
        "topsql.InstanceResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topsql.InstanceItem"
                    }
                }
            }
        },
        "topsql.SummaryByItem": {
            "type": "object",
            "properties": {
                "cpu_time_ms": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cpu_time_ms_sum": {
                    "type": "integer"
                },
                "logical_io_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "logical_io_bytes_sum": {
                    "type": "integer"
                },
                "logical_read_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "logical_read_bytes_sum": {
                    "type": "integer"
                },
                "logical_write_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "logical_write_bytes_sum": {
                    "type": "integer"
                },
                "network_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "network_bytes_sum": {
                    "type": "integer"
                },
                "rocksdb_block_read_count": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "rocksdb_block_read_count_sum": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "timestamp_sec": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "topsql.SummaryItem": {
            "type": "object",
            "properties": {
                "cpu_time_ms": {
                    "type": "integer"
                },
                "duration_per_exec_ms": {
                    "type": "number"
                },
                "exec_count_per_sec": {
                    "type": "number"
                },
                "is_other": {
                    "type": "boolean"
                },
                "logical_io_bytes": {
                    "type": "integer"
                },
                "logical_read_bytes": {
                    "type": "integer"
                },
                "logical_write_bytes": {
                    "type": "integer"
                },
                "network_bytes": {
                    "type": "integer"
                },
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topsql.SummaryPlanItem"
                    }
                },
                "rocksdb_block_read_count": {
                    "type": "integer"
                },
                "scan_indexes_per_sec": {
                    "type": "number"
                },
                "scan_records_per_sec": {
                    "type": "number"
                },
                "sql_digest": {
                    "type": "string"
                },
                "sql_text": {
                    "type": "string"
                }
            }
        },
        "topsql.SummaryPlanItem": {
            "type": "object",
            "properties": {
                "cpu_time_ms": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "duration_per_exec_ms": {
                    "type": "number"
                },
                "exec_count_per_sec": {
                    "type": "number"
                },
                "logical_io_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "logical_read_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "logical_write_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "network_bytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "plan_digest": {
                    "type": "string"
                },
                "plan_text": {
                    "type": "string"
                },
                "rocksdb_block_read_count": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "scan_indexes_per_sec": {
                    "type": "number"
                },
                "scan_records_per_sec": {
                    "type": "number"
                },
                "timestamp_sec": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "topsql.SummaryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topsql.SummaryItem"
                    }
                },
                "data_by": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topsql.SummaryByItem"
                    }
                }
            }
        },
        "topsql.TikvNetworkIoCollectionConfig": {
            "type": "object",
            "properties": {
                "detailed_io_enabled": {
                    "type": "boolean"
                },
                "enable": {
                    "type": "boolean"
                },
                "is_multi_value": {
                    "type": "boolean"
                }
            }
        },
        "topsql.UpdateTikvNetworkIoCollectionRequest": {
            "type": "object",
            "required": [
                "enable"
            ],
            "properties": {
                "detailed_io_enabled": {
                    "type": "boolean"
                },
                "enable": {
                    "type": "boolean"
                }
            }
        },
        "topsql.UpdateTikvNetworkIoCollectionResponse": {
            "type": "object",
            "properties": {
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ErrorResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}